refactor: split sessions.rs into per-agent modules
sessions.rs (1241 lines) was a monolithic file mixing core tmux session
management with agent-specific detection logic for Claude Code, Codex,
and OpenCode. Split into a directory module with dedicated files per agent.

- sessions/mod.rs: core tmux listing, process tree, git info, grouping
- sessions/agents/mod.rs: dispatch + shared helpers (capture_pane, is_numbered_option)
- sessions/agents/claude.rs: Claude Code detection logic
- sessions/agents/codex.rs: Codex detection logic
- sessions/agents/opencode.rs: OpenCode detection logic

Also extracts capture_pane() as a shared helper, eliminating ~20 lines of
duplicated tmux capture-pane boilerplate across the three agent modules.

Adding a new agent now requires only a new file + one match arm.
